services/outgoing-control-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1

# Документооборот
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
aiofiles==23.2.1

# Spell checking
language-tool-python==2.7.1
pyspellchecker==0.7.2

# AI/LLM интеграция
openai==1.3.0
ollama==0.1.7

# Обработка текста
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3

# Валидация и анализ
email-validator==2.1.0

# Логирование и мониторинг
structlog==23.2.0
